+++ docs/zh/guide/other/changelogs/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.4](https://github.com/fast-crud/fast-crud/compare/v1.26.3...v1.26.4) (2025-08-14)
+
+### Performance Improvements
+
+* 如果设置了columnSetShow=false ，则不参与用户的手动排序 ([756f1cf](https://github.com/fast-crud/fast-crud/commit/756f1cfbc02ed31f3ece90248e1477c77ea8a652))
+
+## [1.26.3](https://github.com/fast-crud/fast-crud/compare/v1.26.2...v1.26.3) (2025-08-14)
+
+### Performance Improvements
+
+* 添加对话框支持自动保存草稿，saveDraft ([65f7ece](https://github.com/fast-crud/fast-crud/commit/65f7ecef4012ace86f542338a0ef750dd174d6df))
+* 字段支持动态排序 ([8543a15](https://github.com/fast-crud/fast-crud/commit/8543a15813217f0718975f3cfb9751bcbec3182a))
+* 字段自动设置order，从0开始依次递增 ([0c216ba](https://github.com/fast-crud/fast-crud/commit/0c216ba4de070f8cb1e7eb918b09814e1674a2fa))
+
+## [1.26.2](https://github.com/fast-crud/fast-crud/compare/v1.26.1...v1.26.2) (2025-08-08)
+
+### Bug Fixes
+
+* 修复editable 自定义disabled配置后报错的bug ([2bf01da](https://github.com/fast-crud/fast-crud/commit/2bf01daa18606bba1575976131de79ee065c320d))
+
+## [1.26.1](https://github.com/fast-crud/fast-crud/compare/v1.26.0...v1.26.1) (2025-08-07)
+
+### Bug Fixes
+
+* 修复dict-select naive下不生效的bug ([14965b8](https://github.com/fast-crud/fast-crud/commit/14965b8d21f522a6f9553cce0750a047339c861b))
+* 修复editor-code export name错误的bug ([29dee87](https://github.com/fast-crud/fast-crud/commit/29dee87eb221e05ca31cab8c060bded04b064012))
+
+### Performance Improvements
+
+* dict-select组件重构成setup写法，支持slots ([b307061](https://github.com/fast-crud/fast-crud/commit/b3070613f7e8a791f4adec6bd5d318b9365ed923))
+* element 增加自定义列头示例 ([4aa3f0f](https://github.com/fast-crud/fast-crud/commit/4aa3f0f63bbf6c88c5b4ecc63c0ea476d33df669))
+* naive 增加自定义列头示例， 修复naive rowKey警告 ([7ddc147](https://github.com/fast-crud/fast-crud/commit/7ddc1475033113d7a04b21b26fe9faf6ba491c06))
+
 # [1.26.0](https://github.com/fast-crud/fast-crud/compare/v1.25.13...v1.26.0) (2025-07-28)
 
 ### Bug Fixes