+++ docs/zh/guide/other/changelogs/packages/fast-admin/fs-admin-antdv/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.26.4](https://github.com/fast-crud/fast-crud/compare/v1.26.3...v1.26.4) (2025-08-14)
+
+### Performance Improvements
+
+* 如果设置了columnSetShow=false ，则不参与用户的手动排序 ([756f1cf](https://github.com/fast-crud/fast-crud/commit/756f1cfbc02ed31f3ece90248e1477c77ea8a652))
+
+## [1.26.3](https://github.com/fast-crud/fast-crud/compare/v1.26.2...v1.26.3) (2025-08-14)
+
+**Note:** Version bump only for package @fast-crud/fs-admin-antdv
+
+## [1.26.2](https://github.com/fast-crud/fast-crud/compare/v1.26.1...v1.26.2) (2025-08-08)
+
+**Note:** Version bump only for package @fast-crud/fs-admin-antdv
+
+## [1.26.1](https://github.com/fast-crud/fast-crud/compare/v1.26.0...v1.26.1) (2025-08-07)
+
+**Note:** Version bump only for package @fast-crud/fs-admin-antdv
+
 # [1.26.0](https://github.com/fast-crud/fast-crud/compare/v1.25.13...v1.26.0) (2025-07-28)
 
 **Note:** Version bump only for package @fast-crud/fs-admin-antdv